print: Always cast the error value to int
Avoid burden on the caller to cast the value to int. We will get printf
format warning if, for example, the integer type to be passed in is
greater than int, e.g. on x86-64 we have ssize_t != int, this will
generate silly printf format warnings.

Just always cast it to int to clean it up!

+++ src/teavpn2/print.h
@@ -72,7 +72,7 @@ static inline void set_notice_level(uint8_t level)
 
 
 #define PRERF "(errno=%d) %s"
-#define PREAR(NUM) NUM, strerror(NUM)
+#define PREAR(NUM) ((int)(NUM)), strerror((int)(NUM))
 
 #define pr_err		__pr_error
 #define pr_error	__pr_error
